cisco/requirements.txt
APScheduler==3.9.1
asgiref==3.2.10
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Cython==0.29.23
Django==3.1
djangorestframework==3.13.1
gensim==4.1.2
idna==3.3
joblib==1.1.0
nltk==3.7
numpy==1.22.0
pandas==1.4.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.6.2
requests==2.27.0
scipy==1.8.1
six==1.16.0
smart-open==6.0.0
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
